move config out to another file
